streamlit-dashboard/pyproject.toml
[project]
name = "streamlit-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "db-dtypes>=1.5.1",
    "google-cloud-bigquery>=3.41.0",
    "google-cloud-bigquery-storage>=2.37.0",
    "pandas>=2.3.3",
    "plotly>=6.7.0",
    "streamlit>=1.56.0",
]
